requirements.txt
asgiref==3.3.1
Django==2.2.10
pytz==2020.5
sqlparse==0.4.1
djangorestframework==3.9.2
django-cors-headers==2.5.2
coreapi==2.3.3 
pyyaml==5.1
django-rest-swagger==2.2.0
